requirements.txt
requests
faiss-cpu
numpy
tavily-python
sentence-transformers
python-dotenv
ollama
reportlab
